dynamic schemas for most streams
