--- tap_googleads/schemas/campaign_performance_by_gender_and_device.json
@@ -1,100 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "customer_id": {
-            "type": "string"
-        },
-        "campaign": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "advertisingChannelType": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "adGroup": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "metrics": {
-            "type": "object",
-            "properties": {
-                "clicks": {
-                    "type": "string"
-                },
-                "costMicros": {
-                    "type": "string"
-                },
-                "ctr": {
-                    "type": "number"
-                },
-                "averageCpc": {
-                    "type": "number"
-                },
-                "impressions": {
-                    "type": "string"
-                }
-            }
-        },
-        "adGroupCriterion": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "gender": {
-                    "type": "object",
-                    "properties": {
-                        "type": {
-                            "type": "string"
-                        }
-                    }
-                },
-                "systemServingStatus": {
-                    "type": "string",
-                    "enum": [
-                        "ELIGIBLE",
-                        "RARELY_SERVED",
-                        "UNKNOWN",
-                        "UNSPECIFIED"
-                    ]
-                }
-            }
-        },
-        "genderView": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                }
-            }
-        },
-        "segments": {
-            "type": "object",
-            "properties": {
-                "device": {
-                    "type": "string"
-                },
-                "date": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
