--- tap_googleads/schemas/campaign_performance_by_location.json
@@ -1,74 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "customer_id": {
-            "type": "string"
-        },
-        "campaign": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "metrics": {
-            "type": "object",
-            "properties": {
-                "clicks": {
-                    "type": "string"
-                },
-                "costMicros": {
-                    "type": "string"
-                },
-                "ctr": {
-                    "type": "number"
-                },
-                "averageCpc": {
-                    "type": "number"
-                },
-                "impressions": {
-                    "type": "string"
-                }
-            }
-        },
-        "campaignCriterion": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "location": {
-                    "type": "object",
-                    "properties": {
-                        "geoTargetConstant": {
-                            "type": "string"
-                        },
-                        "bidModifier": {
-                            "type": "number"
-                        }
-                    }
-                }
-            }
-        },
-        "segments": {
-            "type": "object",
-            "properties": {
-                "date": {
-                    "type": "string"
-                }
-            }
-        },
-        "locationView": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
