--- tap_googleads/schemas/click_view_report.json
@@ -1,87 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "type": "object",
-    "properties": {
-        "clickView": {
-            "type": "object",
-            "properties": {
-                "gclid": {
-                    "type": "string"
-                },
-                "adGroupAd": {
-                    "type": "string"
-                },
-                "keyword": {
-                    "type": "string"
-                },
-                "keywordInfo": {
-                    "type": "object",
-                    "properties": {
-                        "matchType": {
-                            "type": "string"
-                        }
-                    }
-                }
-            }
-        },
-        "customer": {
-            "type": "object",
-            "properties": {
-                "id": {
-                    "type": "string"
-                }
-            }
-        },
-        "adGroup": {
-            "type": "object",
-            "properties": {
-                "id": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "campaign": {
-            "type": "object",
-            "properties": {
-                "id": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "date": {
-            "type": "string",
-            "format": "date"
-        },
-        "segments": {
-            "type": "object",
-            "properties": {
-                "adNetworkType": {
-                    "type": "string"
-                },
-                "device": {
-                    "type": "string"
-                },
-                "slot": {
-                    "type": "string"
-                },
-                "clickType": {
-                    "type": "string"
-                }
-            }
-        },
-        "metrics": {
-            "type": "object",
-            "properties": {
-                "clicks": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
