--- tap_googleads/schemas/geo_performance.json
@@ -1,61 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "customer_id": {
-            "type": "string"
-        },
-        "campaign": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            }
-        },
-        "metrics": {
-            "type": "object",
-            "properties": {
-                "clicks": {
-                    "type": "string"
-                },
-                "costMicros": {
-                    "type": "string"
-                },
-                "impressions": {
-                    "type": "string"
-                },
-                "conversions": {
-                    "type": "number"
-                }
-            }
-        },
-        "segments": {
-            "type": "object",
-            "properties": {
-                "date": {
-                    "type": "string"
-                }
-            }
-        },
-        "geographicView": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "locationType": {
-                    "type": "string"
-                },
-                "countryCriterionId": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
