--- tap_googleads/schemas/geo_target_constant.json
@@ -1,34 +1,0 @@
-{
-    "type": "object",
-    "properties": {
-        "customer_id": {
-            "type": "string"
-        },
-        "geoTargetConstant": {
-            "type": "object",
-            "properties": {
-                "resourceName": {
-                    "type": "string"
-                },
-                "status": {
-                    "type": "string"
-                },
-                "id": {
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                },
-                "countryCode": {
-                    "type": "string"
-                },
-                "targetType": {
-                    "type": "string"
-                },
-                "canonicalName": {
-                    "type": "string"
-                }
-            }
-        }
-    }
-}
